try to setItem on episode id
