+++ app/controllers/episodes_controller.rb
@@ -20,6 +20,7 @@ class EpisodesController < ApplicationController
   end
 
   def show
+    @episode = Episode.find(params[:id])
   end
 
   def update
