+++ app/javascript/player/index.js
@@ -1,2 +1,3 @@
 import "./podcast"
 import "./show"
+import "./episode_show"
