+++ app/javascript/player/podcast.js
@@ -1,10 +1,8 @@
 import Rails from "@rails/ujs"
-import Swal from 'sweetalert2'
 
 document.addEventListener("turbolinks:load", function () {
 
   let subsbtn = document.querySelector(".subscribed-btn")
-  let favbtn = document.querySelector(".my_fav")
 
 // 訂閱功能
   if (subsbtn) {
@@ -28,29 +26,5 @@ document.addEventListener("turbolinks:load", function () {
         },
       })
     })
-    
   }
-
-// localstorage 我的最愛
-  function savefav(e){
-    var star = document.querySelector('.my_fav').status;
-      localStorage.setItem('favortie', star);
-      // key = podcast id // 1_2_favortie
-  }
-
-  if (favbtn) {
-    favbtn.addEventListener("click", function (){
-      favbtn.innerHTML = '<i class="fas fa-star"></i>'
-      favbtn.classList.add("text-yellow-300")
-
-        Swal.fire({
-        icon: 'success',
-        title: '已加到我的最愛！',
-        showConfirmButton: false,
-        timer: 1500
-      })
-    })
-  }
-
-
 })
